requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==77.0.3
troposphere==4.9.0
zappa==0.59.0

